server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-telegram-bot==21.10
pyyaml==6.0.2
slowapi==0.1.9
python-dotenv==1.0.1
python-multipart==0.0.22
